Fix-Create-Feature

+++ src/main/java/com/protaskify/protaskify_api/repository/FeatureRepository.java
@@ -4,7 +4,9 @@ import com.protaskify.protaskify_api.model.enity.Feature;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface FeatureRepository extends JpaRepository<Feature,Integer> {
+public interface FeatureRepository extends JpaRepository<Feature, Long> {
 
 }
